refactor(plan_SERVERSPLIT slice 8): the element-extraction bodies leave server.py
The nine `element-extraction` tool bodies move VERBATIM out of
embedded/server.py into embedded/tool_sections/element_extraction.py. Pure move:
the only edits are the dropped registration decorator (contract rule 2) and the
rewrite of the singleton/helper reads to `rt.<name>`, resolved at CALL time
against the one patchable home (contract rule 3).

This is the plan's SECOND stray relocation. `extract_complete_element_cdp` was
physically filed among the file-extraction bodies — wedged between
extract_complete_element_to_file and extract_element_styles_to_file — while
registering into `element-extraction` all along. It lands in its own section's
module, last in its own section's surface order, which is where it registered
from; module <-> section stays a clean 1:1.

Three of these bodies return through response_handler.handle_response, which is
SYNCHRONOUS: awaiting its dict return raises TypeError and silently broke these
very tools once (F-202). This is the slice that proves slice 0's widening of
that guard was load-bearing: tests/test_server_call_conventions.py re-derives
its AST scan from tests/source_scan.py, so it follows the call sites into the
new file instead of passing vacuously over an emptier server.py, and its
`>= 5 call sites` floor still holds.

The three goldens under this section's subsystem
(tests/goldens/extract_element_styles.json, cdp_complete_element.json,
canonical_engine.json) are driven by tests/test_cloner_schemas.py and
tests/test_cdp_element_cloner.py against `cdp_element_cloner` DIRECTLY, one
layer below the tool bodies, so a tool-body move cannot reach them; both files
run in this slice's lane and `git diff -- tests/goldens/` is empty.

server.py 2116 -> 1748 LOC (367 body lines plus the one-line
`cdp_element_cloner` alias import they were the last consumer of); the
GRANDFATHER cap ratchets DOWN to the measured post-ruff-format actual, and
tests/source_scan.py's floor ratchets UP 9 -> 10.

Verified: wire surface IDENTICAL to tests/goldens/tool_surface.json;
SECTION_TOOLS 94 with per-app [94, 94, 94]; `--list-sections` subprocess
Total: 94 tools with element-extraction at 9; unit lane 2330 passed / 1 skipped
(the -1 vs slice 7 is section 5.5's parametrize shrinking by the pruned alias);
real-Chrome test_e2e_data_tools.py 5, test_e2e_hard_dom.py 4,
test_e2e_animations.py 8, test_e2e_animations_edge.py 26 and
test_e2e_transport.py 1, all passed, one file at a time; budgets, ruff
format+check, suppression owners, vulture, ty (81 warnings, 0 errors) and
gen_release_contract --check all clean.
