+++ src/stealth_chrome_devtools_mcp/embedded/tool_sections/__init__.py
@@ -31,6 +31,7 @@ from stealth_chrome_devtools_mcp.embedded.tool_sections import (
     cookies_storage,
     debugging,
     dynamic_hooks,
+    element_extraction,
     file_extraction,
     network_debugging,
     progressive_cloning,
@@ -52,6 +53,7 @@ SECTION_MODULES = (
     cookies_storage,
     debugging,
     tabs,
+    element_extraction,
     progressive_cloning,
     file_extraction,
     dynamic_hooks,
